+++ tui/MaterialDesign/Text.cpp
@@ -1,7 +1,7 @@
 #include "Text.h"
 
 std::vector<char> &Memory::tui::Text::render(bool shouldNotifyParent) {
-    for (int i = 0; i < text.size(); i++) {
+    for (int i = 0; i < (int)text.size(); i++) {
         framebuffer[i] = text[i];
         if (highlighted) framebuffer[i] |= HIGHLIGHT_TEXT;
     }
